+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Unit tests updated to latest glTF-Sample-Models
+
 ## [2.1.0] - 2020-10-25
 ### Changed
 - Updated KTX/Basis Texture Unity Package to 0.8.x